% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex1.R
\name{Ex1}
\alias{Ex1}
\title{Draw barplots for variables}
\usage{
Ex1(data, x, y, na.rm = FALSE)
}
\arguments{
\item{x}{The first variable extracted to represent facets in the graph}

\item{y}{The second variable, which is one of the parameters used to draw the density
graph}
}
\value{
A density graph depicting the relationship between one variable and another
variable in the data set
throws error if 'x' and 'y' are not proper type
}
\description{
Draw barplots for variables
}
\examples{
library(datateachr)
Ex1(cancer_sample, diagnosis, radius_mean)
Ex1(cancer_sample, diagnosis, perimeter_mean)
Throws an error
Ex1(cancer_sample, radius_mean, perimeter_mean)
}
